requirements.txt
Django==1.11.6
django-jet==1.0.6
olefile==0.44
Pillow==4.3.0
pytz==2017.2
django-ckeditor
-e git+https://github.com/mariocesar/sorl-thumbnail.git#egg=sorl-thumbnail
django-modeltranslation
solid_i18n